backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
pydantic>=2.9.0
pandas>=2.2.3
openpyxl>=3.1.5
matplotlib>=3.10.0
fpdf2>=2.8.1
google-genai>=1.0.0
python-dotenv>=1.0.0
pytest>=8.3.0
pytest-mock>=3.14.0
httpx>=0.27.0
